[MINOR] Remove unused TODO comments in CelebornTierProducerAgent#processBuffer
### What changes were proposed in this pull request?
Remove unused TODO comments in CelebornTierProducerAgent#processBuffer

### Why are the changes needed?
In order for buffers to be packed together, we are going to modify the Flink side implementation to delegate buffer compression to tiers. But after discussion, we have been able to handle the case of receiving the compressed buffer on the Celeborn side, so this TODO is no longer needed.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
No need.

Closes #2883 from reswqa/remove_unused_todo.

Authored-by: Weijie Guo <reswqa@163.com>
Signed-off-by: Shuang <lvshuang.xjs@alibaba-inc.com>