[CELEBORN-1504][FOLLOWUP] Start test nodes in random ports to allow multiple builds run in the same ci server
### What changes were proposed in this pull request?

Start test nodes in random ports to allow multiple builds run in the same ci server, which improves the test implementations so that it starts test nodes in random ports instead of using the hardcoded ones.

Follow up #2619.

### Why are the changes needed?

The test nodes are started in the hard coded ports, this prevents to run multiple builds in the same CI/CD server at present.

Bump the changes of #2237.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

`RemoteShuffleMasterSuiteJ`

Closes #2902 from SteNicholas/CELEBORN-1504.

Authored-by: SteNicholas <programgeek@163.com>
Signed-off-by: mingji <fengmingxiao.fmx@alibaba-inc.com>