[CELEBORN-1726] Update WorkerInfo when transition worker state
### What changes were proposed in this pull request?
Update WorkerInfo when transition worker state

### Why are the changes needed?
When we send a `getWorkerInfo` request to the `Worker` node after this worker has became `Decommission` state , it does not return the correct state.

note: If we send this request to `Master` node instead of `Worker`, the return value is correct. We should update the worker state also for worker node itself.

### Does this PR introduce _any_ user-facing change?
Yes

### How was this patch tested?
Mannually

Closes #2930 from reswqa/worker_info.

Authored-by: Weijie Guo <reswqa@163.com>
Signed-off-by: Shuang <lvshuang.xjs@alibaba-inc.com>