[CELEBORN-1749] Fix incorrect application diskBytesWritten metrics
### What changes were proposed in this pull request?
Fix the incorrect application diskBytesWritten metrics, also make some code refactor to easy understand the logic.

### Why are the changes needed?

There is typo, before when updating the top app usage metrics, it removes the gauge from `workerSrouce`, then the gauge in the `resourceConsumptionSource` will not be refreshed, and the old metrics would be never cleaned.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
UT

Before this PR:
The application metrics is incorrect. The application resource consumption is 0 and will not update.
<img width="1917" alt="image" src="https://github.com/user-attachments/assets/035b9590-1b31-4d58-af0f-ce2fd3e71d0e">

After this PR:
The application `diskBytesWritten` equals the user `diskBytesWritten` and update with the app consumption change, it is expected.

<img width="1918" alt="image" src="https://github.com/user-attachments/assets/11bf6037-f696-4ab1-9685-f89f4dd1c692">

Closes #2954 from turboFei/fix_metrics_wrong.

Authored-by: Wang, Fei <fwang12@ebay.com>
Signed-off-by: SteNicholas <programgeek@163.com>