[CELEBORN-1750] Return struct worker resource consumption information with RESTful api
### What changes were proposed in this pull request?

As title

### Why are the changes needed?

This information is useful for the automation tool integration.

For our automation tool, it query all the worker status periodically by calling master `/api/v1/worker`.

In decommission process, when the worker is in IDLE state, we need to check whether there is still unreleased shuffle data on this worker so that we can shutdown this node without user impaction.

Before, I have to call the worker `/ap1/v1/shuffles` to check that.

It is better that we can get all the information from celeborn master end, because master is HA enabled and always reachable.

So in this PR, it returns the struct resource consumption for automation tool integration.

### Does this PR introduce _any_ user-facing change?

No, this RESTful api has not been released.

### How was this patch tested?
GA.

Closes #2955 from turboFei/worker_info_object.

Authored-by: Wang, Fei <fwang12@ebay.com>
Signed-off-by: Wang, Fei <fwang12@ebay.com>