[CELEBORN-1670] Avoid swallowing InterruptedException in ShuffleClientImpl
### What changes were proposed in this pull request?
As title.

### Why are the changes needed?
In the ShuffleClientImpl, methods such as pushData and pushMergedData might encounter interruptions during message transmission via the TransportClient. However, the InterruptedException may be ignored, as it is handled as a standard exception. As a result, the ShuffleClientImpl continues its operation(retry or revive) even when an InterruptedException occurs.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Add UT: org.apache.celeborn.client.ShuffleClientSuiteJ#testPushDataAndInterrupted

Closes #2849 from jiang13021/interrupt_in_shuffle_client.

Authored-by: jiang13021 <jiangyanze.jyz@antgroup.com>
Signed-off-by: Shuang <lvshuang.xjs@alibaba-inc.com>