[CELEBORN-1510] Partial task unable to switch to the replica
### What changes were proposed in this pull request?

Attempt%2 = 1 unable to switch to the replica

### Why are the changes needed?

When Attempt equals 1, createReader will access location.peer. Special circumstances can lead to unexpected behaviors. For example, an exception occurs during the process of obtaining data and the peer needs to be used. However, this logic will switch to the abnormal node again.

<img width="1626" alt="image" src="https://github.com/user-attachments/assets/21c50953-db0f-4717-9b91-e3aeae16ece2">
<img width="1652" alt="image" src="https://github.com/user-attachments/assets/7430786c-26a4-4b3b-be68-f8bdf780c58c">

### Does this PR introduce _any_ user-facing change?
NO

### How was this patch tested?
Internal test.

Closes #2626 from zhaostu4/switch_replica.

Authored-by: zhangzhao.08 <zhangzhao.08@bytedance.com>
Signed-off-by: Shuang <lvshuang.xjs@alibaba-inc.com>