[CELEBORN-1794] Fix TestCongestionController occasional failing
### Why are the changes needed?
There are a small probability of the TestCongestionController test failing.

![image](https://github.com/user-attachments/assets/bc5bfb91-0b40-4ee0-bd74-2b96715d7cd7)

That is because the `checkService` will excute once it was init, which can cause a multithreading conflict with the test code.

![image](https://github.com/user-attachments/assets/dff58b08-a340-4c29-a61d-fafaeb182c5d)

### What changes were proposed in this pull request?
Fix ut bug.
In fact, `shutDownCheckService` still wont prevent the `checkService` from excuting at once but can make the main testing thread waiting for it to shutDown.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
manual test.

Closes #3017 from zaynt4606/clb1794.

Authored-by: zhengtao <shuaizhentao.szt@alibaba-inc.com>
Signed-off-by: Shuang <lvshuang.xjs@alibaba-inc.com>