[CELEBORN-1846] Fix the StreamHandler usage in fetching chunk when task attempt is odd
### What changes were proposed in this pull request?
The streams opened in the `streamCreatorPool` thread pool are all based on the primary locations. When the task attempt is odd, the task will start to fetch the chunk from the replica location first. This will cause using the wrong streamHandler to fetch data. To keep the logic simple, we always fetch from the primary location, and when change to peer, closing stream and use a null streamHandler when fetching peers.

### Why are the changes needed?
Avoid tasks that are slowed down by NPE and potential data problems.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Manual test.

Closes #3079 from onebox-li/fix-stream-handler.

Authored-by: onebox-li <lyh-36@163.com>
Signed-off-by: Shuang <lvshuang.xjs@alibaba-inc.com>