[CELEBORN-1867][0.5] Fix flink client memory leak of TransportResponseHandler#outstandingRpcs for handling addCredit response
### What changes were proposed in this pull request?

Backport https://github.com/apache/celeborn/pull/3103 to branch-0.5.

When I tested the performance of Flink with Celeborn in session mode, using both the shuffle-service plugin and hybrid shuffle integration strategies, I noticed that the task heap continuously grew even when no jobs were running.

The issue arises because the Celeborn client sends addCredit or notifyRequiredSegment requests, expecting a response. This creates a callback and maintains a reference to CelebornBufferStream, SingleInputGate, and StreamTask.
These callbacks are stored in TransportResponseHandler#outstandingRpcs and are cleared upon receiving a response.

However, the Worker does not send a response for these two RPC requests, which leads to a significant memory leak in the client.

### Why are the changes needed?

It may cause flink client memory leak.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Test by rerun TPCDS in Flink session mode.

Closes #3123 from SteNicholas/CELEBORN-1867.

Authored-by: codenohup <huangxu.walker@gmail.com>
Signed-off-by: Shuang <lvshuang.xjs@alibaba-inc.com>