[CELEBORN-1841][0.5] Support custom implementation of EventExecutorChooser to avoid deadlock when calling await in EventLoop thread
### What changes were proposed in this pull request?

Support custom implementation of `EventExecutorChooser` to avoid deadlock when calling await in EventLoop thread.

Backport: #3071.

### Why are the changes needed?

In Flink Celeborn Client, you can create a new connection in the EventLoop thread. To wait for the connection to complete, cf.await is called, which can cause a deadlock because the thread bound to the newly connected channel may be the same as the current EventLoop thread. The current thread is suspended by wait and needs to wait for the current thread to notify. This change is to avoid binding the same thread.

### Does this PR introduce _any_ user-facing change?

`celeborn.<module>.io.conflictAvoidChooser.enable` is introduced.

### How was this patch tested?

Manual test.

Closes #3140 from SteNicholas/CELEBORN-1841.

Authored-by: xinyuwang1 <xinyuwang1@xiaohongshu.com>
Signed-off-by: SteNicholas <programgeek@163.com>