[CELEBORN-1844][CIP-8] introduce tier writer proxy and simplify partition data writer
### What changes were proposed in this pull request?
Most of the contents of this PR is the changes for UT, and it's derived from another PR #3083.

### Why are the changes needed?
Introduce tier writer proxy and simplify the partition data writer.

### Does this PR introduce _any_ user-facing change?
NO.

### How was this patch tested?
GA and manual tests.
I have verified data correctness on a cluster using TPC-DS tests.
The average performance tests result for this PR:
with this PR : without this PR =  2562.5 : 2576.85
There is no performance reduction.

Closes #3085 from FMX/1844.

Authored-by: mingji <fengmingxiao.fmx@alibaba-inc.com>
Signed-off-by: Shuang <lvshuang.xjs@alibaba-inc.com>