[CELEBORN-1966] Added fix to get userinfo in celeborn
### What changes were proposed in this pull request?
Updated the DefaultIdentityProvider to return tenantId, name based on configs provided by the user. Using the CelebornConf object used by LifeCycleManager to get the tenantId and name.

### Why are the changes needed?
The tenant id and username passed by the user were not being used because the DefaultIdentityProvider creates a new CelebornConf each time the provide function is called. Due to this, the tenantid and username were always coming as default. With these changes, we are using the CelebornConf object used by the LifeCycleManager which includes the configs provided by the user.

### Does this PR introduce _any_ user-facing change?
NA

### How was this patch tested?
Added UTs and tested on staging setup

Closes #3214 from AmandeepSingh285/CELEBORN-1966.

Authored-by: amandeeps.28 <amandeeps.28@uber.com>
Signed-off-by: Shuang <lvshuang.xjs@alibaba-inc.com>