[CELEBORN-1948] Fix the issue where replica may lose data when HARD_SPLIT occurs during handlePushMergeData
### What changes were proposed in this pull request?
Fix data lost when push merged data of replica and hard split happen.

### Why are the changes needed?
There is a problem with replicate rpc callback. The code should satisfy the following conditions: when comparing the status returned by primary and replica data, the status on the left should be used as the final status for the client, FAILURE > HARD_SPLIT > CONGESTION > SUCCESS. The status on the right cannot cover the status on the left.

There are two problems with the code now
1. HARD_SPLIT can cover FAILURE, which will affect the exclude worker logic, and there may be some problems
2. When processing a pushMergedData request, some partitionLocations are committed, PushDataHandler cannot stop pushing replicas as long as there are any partitions that have not been committed, otherwise data loss will occur.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Existing UTs.

Closes #3185 from leixm/CELEBORN-1948.

Lead-authored-by: Xianming Lei <31424839+leixm@users.noreply.github.com>
Co-authored-by: Xianming Lei <xianming.lei@shopee.com>
Co-authored-by: Shuang <lvshuang.xjs@alibaba-inc.com>
Signed-off-by: Shuang <lvshuang.xjs@alibaba-inc.com>