[CELEBORN-1974] ApplicationId as metrics label should be behind a config flag
### What changes were proposed in this pull request?

Push applicationId as metrics label only if `celeborn.metrics.worker.appLevel.enabled` is true.

### Why are the changes needed?

At Uber, We use m3 for monitoring, it tries to make a new series using all the present metrics label. Having applicationId as a metrics introduces too much cardinality in `activeconnectioncount` and we are unable to use it, while it is an useful metric with/without applicationId as label. Similarly for resourceConsumption, userIdentifier alone can be used.

### Does this PR introduce _any_ user-facing change?
Yes, changed the default config value.

### How was this patch tested?
NA

Closes #3221 from s0nskar/application_tag.

Lead-authored-by: Sanskar Modi <sanskarmodi97@gmail.com>
Co-authored-by: Wang, Fei <fwang12@ebay.com>
Signed-off-by: Wang, Fei <fwang12@ebay.com>