[CELEBORN-2009] Commit files request failure should exclude worker in LifecycleManager
### What changes were proposed in this pull request?

Exclude worker in lifecycle manager if the commit files request on workers fails with `COMMIT_FILE_EXCEPTION` or after multiple retries.

### Why are the changes needed?

If worker is under high load and not able to process request because of high CPU, we should exclude it so it will not affect the next retry to shuffle stage.

Internally, we are seeing commit file futures in worker under high load are getting timed out and next retry of the stage is again picking same servers and failing. Similarly, we are seeing continuous RpcTimeout for workers but those workers are again getting selected for next retry.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?

NA

Closes #3276 from s0nskar/worker_exlude_on_commit_exception.

Authored-by: Sanskar Modi <sanskarmodi97@gmail.com>
Signed-off-by: Wang, Fei <fwang12@ebay.com>
(cherry picked from commit aeac31f6f062db838566c9cbc8649b48f97db298)
Signed-off-by: Wang, Fei <fwang12@ebay.com>