[CELEBORN-2160] Speedup CommitHandler.finishMapperAttempt
### What changes were proposed in this pull request?
Speedup `finishMapperAttempt` by making its complexity constant instead of linear to number of mappers

### Why are the changes needed?
As detailed in [CELEBORN-2160](https://issues.apache.org/jira/browse/CELEBORN-2160), when there are a large number of concurrent 'mapper' tasks for a stage with a large mapper partitions - the linear (to number of mappers) complexity of `finishMapperAttempt` results in rpc timing out and failing.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing tests.

Closes #3487 from mridulm/speedup-finish-mapper-attempts.

Authored-by: Mridul Muralidharan <mridulatgmail.com>
Signed-off-by: Mridul Muralidharan <mridul<at>gmail.com>